dataloader/requirements.txt
requests
google-cloud-storage
python-dotenv